requirements.txt
# Core dependencies
mem0ai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MCP dependencies
mcp>=0.1.0

# Async support
asyncio
aiohttp>=3.9.0

# Logging and monitoring
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0